requirements.txt
python-doctr[torch]
opencv-python
numpy<2.0.0
streamlit
pillow
openai
anthropic
python-dotenv
pytesseract
transformers
imutils
